reorg folders

+++ packages/fetch/src/fetch.py
@@ -4,16 +4,13 @@ from alpha_vantage.fundamentaldata import FundamentalData
 
 from polygon import RESTClient
 
-# from datapackage import Package
-
-# package = Package('https://datahub.io/core/finance-vix/datapackage.json')
-
 import numpy as np
 import pandas as pd
 from datetime import datetime as dt
 from time import sleep
 
-from pandas import DataFrame
+from pandas import DataFrame, Timestamp
+from typing import Any, Optional
 
 """
 REFERENCES:
@@ -25,20 +22,14 @@ REFERENCES:
     
 """
 
-class Fetch_Alpha:
-    
-    def __init__(self,tickers:list[str],startDate:pd.Timestamp,endDate:pd.Timestamp=None):
-        self.key = 'WKSR42529JVOYXM9'
-        
-        self.ts = TimeSeries(self.key, output_format='pandas')
-        self.ti = TechIndicators(self.key)
-        self.fd = FundamentalData(self.key)
-        
+class Fetch:
+
+    def __init__(self,tickers:list[str],startDate:Timestamp,endDate:Optional[Timestamp]=None):
         self.tickers = tickers
         self.startDate = startDate
         
         if endDate is None:
-            self.endDate = pd.Timestamp(year=dt.now().year,month=dt.now().month,day=dt.now().day)
+            self.endDate = Timestamp(year=dt.now().year,month=dt.now().month,day=dt.now().day)
         else:
             self.endDate = endDate
 
@@ -47,10 +38,12 @@ class Fetch_Alpha:
             Data = self.getPrices()
 
             if writeOpt:
-                with pd.ExcelWriter('inputs/data.xlsx') as writer:
-                    for ticker in self.tickers:
-                        data = Data[ticker]
-                        data.to_excel(writer, sheet_name=ticker)
+                cont = bool(input("Are you sure you'd like to overwrite data? (0/1)"))
+                if cont:
+                    with pd.ExcelWriter('inputs/data.xlsx') as writer:
+                        for ticker in self.tickers:
+                            data = Data[ticker]
+                            data.to_excel(writer, sheet_name=ticker)
         else:
             Data = {}
             for ticker in self.tickers:
@@ -60,9 +53,23 @@ class Fetch_Alpha:
                     Data[ticker] = data
 
         return Data
+    
+    def getPrices(self) -> dict[str,DataFrame]:
+        pass
+
+class Fetch_Alpha(Fetch):
+    
+    def __init__(self,tickers:list[str],startDate:Timestamp,endDate:Optional[Timestamp]=None):
+        self.key = 'WKSR42529JVOYXM9'
         
-    def getPrices(self,increment='daily',splitOpt=False,divOpt=False):
-        data = {}
+        self.ts = TimeSeries(self.key, output_format='pandas')
+        self.ti = TechIndicators(self.key)
+        self.fd = FundamentalData(self.key)
+
+        super().__init__(tickers,startDate,endDate)
+        
+    def getPrices(self,increment='daily',splitOpt=False,divOpt=False) -> dict[str,DataFrame]:
+        data:dict[str,DataFrame] = {}
         for ticker in self.tickers:
             while True:
                 # data[ticker],_ = self.ts.get_daily_adjusted(symbol=ticker,outputsize='full')
@@ -104,103 +111,50 @@ class Fetch_Alpha:
                     div += data[ticker]['Dividend'].loc[date]
             
         return data
-    
+
+################################################################################   
+
 """
 REFERENCES:
     
     
 """
 
-class Fetch_Polygon:
+class Fetch_Polygon(Fetch):
     
-    def __init__(self):#,tickers:list[str],startDate:pd.Timestamp,endDate:pd.Timestamp=None):
+    def __init__(self,tickers:list[str],startDate:Timestamp,endDate:Optional[Timestamp]=None):
         self.key = 'ez7NxSEphpNlHZFjsp6MSANJxs1UNOGv'
         
-        client = RESTClient(self.key)
-
-        aggs = []
-        for a in client.list_aggs(
-            "AAPL",
-            1,
-            "minute",
-            "2022-01-01",
-            "2023-02-03",
-            limit=50000,
-        ):
-            aggs.append(a)
-
-        print(aggs)
+        self.client = RESTClient(self.key)
         
-        # self.tickers = tickers
-        # self.startDate = startDate
+        super().__init__(tickers,startDate,endDate)
         
-        # if endDate is None:
-        #     self.endDate = pd.Timestamp(year=dt.now().year,month=dt.now().month,day=dt.now().day)
-        # else:
-        #     self.endDate = endDate
-
-    def fetch(self, fetchOpt:bool=False, writeOpt:bool=False) -> dict[str,DataFrame]:
-        if fetchOpt:
-            Data = self.getPrices()
-
-            if writeOpt:
-                with pd.ExcelWriter('inputs/data.xlsx') as writer:
-                    for ticker in self.tickers:
-                        data = Data[ticker]
-                        data.to_excel(writer, sheet_name=ticker)
-        else:
-            Data = {}
-            for ticker in self.tickers:
-                if not fetchOpt:
-                    data = pd.read_excel('inputs/data.xlsx', sheet_name=ticker, index_col=0)
-                    data = data.loc[self.startDate:self.endDate]
-                    Data[ticker] = data
-
-        return Data
-        
-    def getPrices(self,increment='daily',splitOpt=False,divOpt=False):
-        data = {}
+    def getPrices(self,increment:str='day') -> dict[str,DataFrame]:
+        data:dict[str,DataFrame] = {}
         for ticker in self.tickers:
             while True:
                 # data[ticker],_ = self.ts.get_daily_adjusted(symbol=ticker,outputsize='full')
                 try:
-                    if increment == 'daily':
-                        data[ticker],_ = self.ts.get_daily(symbol=ticker,outputsize='full')
-                    elif increment == 'weekly':
-                        data[ticker],_ = self.ts.get_weekly(symbol=ticker)
-                    elif increment == 'monthly':
-                        data[ticker],_ = self.ts.get_monthly(symbol=ticker)
+                    aggs = self.client.get_aggs(
+                        ticker=ticker,
+                        multiplier=1,
+                        timespan=increment,
+                        from_=self.startDate,
+                        to=self.endDate,
+                        adjusted=True
+                    )
+                    
+                    data[ticker] = pd.DataFrame([vars(a) for a in aggs],columns=list(vars(aggs[0]).keys()))
+                    data[ticker]['Date'] = pd.Series([dt.fromtimestamp(timestamp) for timestamp in data[ticker].loc[:,'timestamp']/1000])
+                    data[ticker].drop(['timestamp','transactions','otc','vwap'], axis=1, inplace=True) 
 
                     print('Fetched: ' + ticker + '...')
                     break
                 except ValueError:
                     print('WARNING: Call limit per minute exceeded, waiting 5 seconds...')
                     sleep(5)
-            
-            data[ticker].index.name = 'Date'
-            # data[ticker].columns = ['Open','High','Low','Close','Adjusted','Volume','Dividend','Split']
+
+            data[ticker].set_index('Date',inplace=True)
             data[ticker].columns = ['Open','High','Low','Close','Volume']
-            data[ticker] = data[ticker].reindex(index=data[ticker].index[::-1])
-            
-            data[ticker] = data[ticker].loc[self.startDate:self.endDate]
-            
-            if splitOpt:
-                split = 1
-                for date in data[ticker].index.values[::-1]:
-                    for key in ['Open','High','Low','Close']:
-                        data[ticker][key].loc[date] /= split
-                        
-                    split *= data[ticker]['Split'].loc[date]
-                    
-            if divOpt:
-                div = 0
-                for date in data[ticker].index.values[::-1]:
-                    for key in ['Open','High','Low','Close']:
-                        data[ticker][key].loc[date] -= div
-                        
-                    div += data[ticker]['Dividend'].loc[date]
             
         return data
-    
-
-fetch = Fetch_Polygon()
